TorisPlantData/requirements.txt
altair==4.1.0
altgraph==0.17
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
bokeh==2.3.2
boto3==1.14.49
botocore==1.17.49
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
cryptography==3.0
cvxpy==1.0.31
cycler==0.10.0
dateutils==0.6.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
Django==3.2.5
django-adaptors==0.2.5
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-v5==1.0.4
django-bootstrap4==3.0.1
django-cors-headers==3.4.0
django-crispy-forms==1.12.0
django-filter==2.3.0
django-select2==7.7.1
django-session-timeout==0.1.0
django-tables2==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
ecdsa==0.14.1
ecos==2.0.7.post1
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.0.1
fastapi==0.66.0
fontawesome==5.10.1.post1
fpdf==1.7.2
future==0.18.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gspread==3.6.0
gspread-dataframe==3.0.7
gspread-pandas==2.2.3
h11==0.12.0
httpie==2.2.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.10.1
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
more-itertools==8.4.0
MouseInfo==0.1.3
multiprocess==0.70.10
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.2.4
notebook==6.1.3
nsepy==0.8
nsetools==1.0.11
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.4
osqp==0.6.1
packaging==20.4
pandas==1.1.0
pandas-bokeh==0.5.5
pandas-datareader==0.9.0
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pdfkit==0.6.1
pefile==2021.5.24
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
preprocessing==0.1.13
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.12.4
psycopg2==2.8.5
py==1.9.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pydantic==1.8.2
pydeck==0.5.0b1
PyGetWindow==0.0.9
Pygments==2.6.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
PyJWT==1.7.1
Pympler==0.8
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
PyPDF3==1.0.1
pyperclip==1.8.1
pypiwin32==223
pyportfolioopt==1.2.4
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
pytest==6.0.1
pytest-timeit==0.3.0
python-dateutil==2.8.1
python-jose==3.2.0
python-vuejs==1.0.9
pyttsx3==2.90
pytube==10.8.5
PyTweening==1.0.3
pytz==2020.1
pywhatkit==3.8
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.2
reportlab==3.5.67
requests==2.24.0
requests-oauthlib==1.3.0
rest-pandas==1.1.0
rsa==4.6
s3transfer==0.3.3
scipy==1.5.2
scs==2.1.2
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.3
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sphinx-rtd-theme==0.2.4
SQLAlchemy==1.3.19
sqlparse==0.3.1
starlette==0.14.2
tablib==3.0.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.54.1
traitlets==4.3.3
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
uvicorn==0.14.0
validators==0.18.0
watchdog==0.10.3
wcwidth==0.2.5
webbot==0.34
webencodings==0.5.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
xlrd==1.2.0
XlsxWriter==1.3.3
yahoo-finance==1.4.0
yahoofinancials==1.6
yfinance==0.1.63
zipp==3.5.0
